requirements.txt
blinker==1.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
Django==2.2.5
django-ckeditor==5.8.0
django-crispy-forms==1.8.1
django-js-asset==1.2.2
idna==2.8
Jinja2==2.10.3
MarkupSafe==1.1.1
oauthlib==3.0.1
psycopg2==2.8.4
pycparser==2.19
PyJWT==1.7.1
pyOpenSSL==19.1.0
PySocks==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.13.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.7
